requirements.txt
# =========================================
# 🚀 InvoAI User Management API Requirements
# =========================================

# --- FastAPI Core ---
fastapi==0.115.0
uvicorn[standard]==0.30.6

# --- Database & ORM ---
SQLAlchemy==2.0.34
alembic==1.14.0
psycopg2-binary==2.9.10

# --- Environment & Config ---
python-dotenv==1.0.1
pydantic==2.8.2
pydantic-settings==2.5.2

# --- Authentication & Security ---
bcrypt==4.1.2
passlib==1.7.4

# --- HTTP Client (Optional for external APIs) ---
httpx==0.27.2

# --- Logging & Observability ---
structlog==24.1.0
python-json-logger==2.0.7

# --- Optional Flask UI (If integrated) ---
Flask==3.0.3
Jinja2==3.1.4

# --- Development Tools ---
black==24.10.0
isort==5.13.2
